font-size property, let absolute-size and relative-size values have larger effect
Using roughly 1.2 per step, a value mentioned in the CSS2.1 spec. Firefox
appears to use a value that is similar but a little larger.
http://css-tricks.com/2580-css-font-size/ says that "keyword sizing is
pretty consistent across browsers and platforms", and perhaps someday
we'll want to go out of our way to make them exactly the same, but this
is a start.