Avoid scroll flickering with a custom tooltip class (TipWin) & a custom button
  Removed  the  workaround  [1] that added a full UI and viewport
redraw  to  conceal  the  lingering tooltips bug [2]. It produced
annoying  scroll  flickering  when  the  mouse  was in the UI but
outside the viewport (e.g. over the panel).

  This  is more a FLTK bug than anything; the FLTK team's plan is
to fix it in FLTK3, and also extend the tooltips API.

  As  FLTK3 may take long to be released, and porting dillo to it
is  non-trivial (if both events ever come to happen), this custom
solution looks quite reasonable for the present times.

  The  patch  introduces a new CustButton class that uses TipWin,
and  thus the tooltip handling is no longer handled by FLTK. This
patch switches the UI buttons to the new CustButton.


1] http://lists.auriga.wearlab.de/pipermail/dillo-dev/2011-July/008515.html
2] http://lists.auriga.wearlab.de/pipermail/dillo-dev/2011-July/008494.html