Fix a bug with injected HTML that could lead to reads out of bounds.
Problem details are in getCurTagLineNumber.html.asan file.
This solution adds a flag to avoid potential HTML bug messages generation for
injected HTML. The problem is that injected HTML lies in a separate buffer
so the line number counter gets lost. BTW, there's no point in the bug message
because the user never sees the browser-injected HTML sources.

In this particular case there was a bug in the original html file that
went unreported. This patch also fixes this.